feat: 🎸 refactor components
