+++ src/apps/Counter/components/Modals/Calculator/CalButton.tsx
@@ -1,21 +1,24 @@
 import { Button } from '@/components/common';
 
 type Props = {
-  text: string
+  text?: string
   onClick: () => void
+  children?: React.ReactNode
 }
 
 export const CalButton = (props: Props) => {
-  const { text, onClick } = props;
+  const { text, onClick, children } = props;
 
   return (
     <Button
-      className='w-full py-2'
+      className='w-full'
       color='primary'
       outline
       onClick={onClick}
     >
       {text}
+
+      {children}
     </Button>
   );
 };
